Consolidating path helpers into a utils module
